#94 [RENAME] SearchPage를 PostSearchPage로 변경
