+++ src/pages/PostPage/PostSearchPage/index.js
@@ -1,0 +1,1 @@
+export { default as PostSearchPage } from './PostSearchPage';
