--- src/pages/SearchPage/SearchPage.module.css
@@ -1,30 +1,0 @@
-.container {
-  width: 100%;
-}
-
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 8px;
-}
-
-.posts {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 8px;
-  /* 양옆 여백 */
-  padding: 4px 17px 110px 17px;
-  row-gap: 8px;
-}
-
-.noResult {
-  padding: 50px 0;
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 150%;
-  letter-spacing: -0.5px;
-}
