--- src/pages/SearchPage/index.js
@@ -1,1 +1,0 @@
-export { default as SearchPage } from './SearchPage';
